// src/app/core/api-v1/model/authResponse.ts
/**
 * SpotCamp ID - Campsite Booking Platform API
 *
 * Contact: api@spotcamp.id
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserInfoDto } from './userInfoDto';


export interface AuthResponse { 
    accessToken?: string;
    expiresIn?: number;
    user?: UserInfoDto;
}